Migrate to six and get rid of 2to3.
Now we've got really a single code base for both Python 2 and 3.
For now it just adds a dependency to six which should be available
on most systems. However, if it turns out to switch back to
no dependencies it should be fairly easy to replace six by
some conditional imports.